Overwrite img src with data-src
Many website (e.g. Wechat) sets a placeholder for img src and loads the actual URL lazily. Mozilla Readability attempted to recognize and handle it, but the logic is narrow and does not cover many cases. Instead, this code change aggressively assume that we can overwrite img src with data-src, etc.